[Feat] delete 버튼클릭시 이미지 업로드 input 초기화 #102
- props drilling 이용
